src/test/java/net/learning/computerscience/datastructures/topic_6_trees/ideserve/p11_depth_of_deepest_odd_level_leaf_node/SolutionTest.java
package net.learning.computerscience.datastructures.topic_6_trees.ideserve.p11_depth_of_deepest_odd_level_leaf_node;

import org.junit.Test;

import static org.hamcrest.CoreMatchers.equalTo;
import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;

public class SolutionTest {

    @Test
    public void getDeepestOddLevelLeafNode() {

        TreeNode n1 = new TreeNode(1);
        TreeNode n2 = new TreeNode(2);
        TreeNode n3 = new TreeNode(3);
        TreeNode n4 = new TreeNode(4);
        TreeNode n5 = new TreeNode(5);
        TreeNode n6 = new TreeNode(6);
        TreeNode n7 = new TreeNode(7);
        TreeNode n8 = new TreeNode(8);
        TreeNode n9 = new TreeNode(9);
        TreeNode n10 = new TreeNode(10);

        n1.left = n2;
        n1.right = n3;
        n3.left = n4;
        n3.right = n5;
        n4.left = n6;
        n5.right = n7;
        n6.right = n8;
        n7.right = n9;
        n9.left = n10;

        assertThat(Solution.getDeepestOddLevelLeafNode(n1), is(equalTo(5)));
    }

    @Test
    public void getDeepestOddLevelLeafNode2() {

        TreeNode n1 = new TreeNode(1);
        TreeNode n2 = new TreeNode(2);
        TreeNode n3 = new TreeNode(3);
        TreeNode n4 = new TreeNode(4);
        TreeNode n5 = new TreeNode(5);
        TreeNode n6 = new TreeNode(6);
        TreeNode n8 = new TreeNode(8);
        TreeNode n9 = new TreeNode(9);

        n1.left = n2;
        n1.right = n3;
        n2.left = n4;
        n2.right = n5;
        n3.left = n6;
        n4.right = n8;
        n6.right = n9;

        assertThat(Solution.getDeepestOddLevelLeafNode(n1), is(equalTo(3)));
    }

}